requirements.txt
jinja2
pyyaml
markupsafe